update audio playing capabilitys to signals current state

+++ src/org/thoughtcrime/securesms/mms/AttachmentManager.java
@@ -271,7 +271,7 @@ public class AttachmentManager {
               public void onStop() {}
 
               @Override
-              public void onProgress(double progress, int millis) {}
+              public void onProgress(double progress, long millis) {}
 
               @Override
               public void onReceivedDuration(int millis) {
